requirements.txt
pytorch
torchvision
opencv-python
gdown
dlib
pandas
numpy